Update minimap.cpp
* Exploration mode:
By default OTClient is drawing all tiles which are send by the server to the minimap view. Some server admins might not want this and will receive a working patch for this. Simply add EXPLORE_MODE to build dependencies to make it work. While still all tiles are received from the server, only the visible ones on the same floor as the player including tiles inside the view range will be drawn. All other tiles will be added to a caching system and drawn later in case the player walks up or down a floor. To prevent the Client from freezing or slowing done after a long play time, the cache will be flushed of all tiles which have been drawn already or which are far out from our view range.

* Fixed an issue inside the code causing the Client to crash
* Made some code cleanup