Last fix + separation of connectors
I knew that it was the second case that was bad but only after checking the history log I was able to identify that I was iterating only from lowest to highest even in the case where I only needed to update the indexes below "index".

Also I was able to detect that some events were conflicting with onCreatureDisappear. This also happened with distance.

Lastly I'm separating the connectors in order for in the future we can connect and disconnect based upon sortType and only use the events which are useful at the moment (and perhaps disabling them all when the module window is closed).